requirements.txt
cv2PySide6==3.1.0
face_recognition==1.3.0
mediapipe==0.9.1.0
Pillow==9.5.0
qimage2ndarray==1.10.0
